+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.9.0
+- **FEAT**(android, iOS, macOS): Add `gpsCoordinates` to `VideoMetadata` for extracting GPS location from videos.
+- **FEAT**(android, iOS, macOS): Add `frameRate` to `VideoMetadata` for extracting video frame rate.
+- **FEAT**(iOS, macOS): Add `cameraMake` and `cameraModel` to `VideoMetadata` for extracting camera information.
+
 ## 1.8.0
 - **FEAT**(android, iOS, macOS): Add `customAudioStartTime` option to `VideoRenderData`. Allows starting the custom audio track from a specific offset instead of from the beginning. Useful for using a specific section of a longer audio file.